Implement TensorflowHamiltonian
