+++ src/qibo/hamiltonians.py
@@ -1,185 +1,11 @@
 # -*- coding: utf-8 -*-
 import numpy as np
 from qibo import matrices, K
-from qibo.config import DTYPES
-
-
-NUMERIC_TYPES = (np.int, np.float, np.complex,
-                 np.int32, np.int64, np.float32,
-                 np.float64, np.complex64, np.complex128)
-
-
-class Hamiltonian(object):
-    """This class implements the abstract Hamiltonian operator.
-
-    Args:
-        nqubits (int): number of quantum bits.
-        matrix (np.ndarray): Matrix representation of the Hamiltonian in the
-            computational basis as an array of shape
-            ``(2 ** nqubits, 2 ** nqubits)``.
-    """
-
-    def __init__(self, nqubits, matrix):
-        if not isinstance(nqubits, int):
-            raise RuntimeError(f'nqubits must be an integer')
-        shape = tuple(matrix.shape)
-        if shape != 2 * (2 ** nqubits,):
-            raise ValueError(f"The Hamiltonian is defined for {nqubits} qubits "
-                              "while the given matrix has shape {shape}.")
-
-        self.nqubits = nqubits
-        self.matrix = matrix
-        self._eigenvalues = None
-        self._eigenvectors = None
-        self._exp = {"a": None, "result": None}
-
-    def eigenvalues(self):
-        """Computes the eigenvalues for the Hamiltonian."""
-        if self._eigenvalues is None:
-            self._eigenvalues = K.linalg.eigvalsh(self.matrix)
-        return self._eigenvalues
-
-    def eigenvectors(self):
-        """Computes a tensor with the eigenvectors for the Hamiltonian."""
-        if self._eigenvectors is None:
-            self._eigenvalues, self._eigenvectors = K.linalg.eigh(
-                self.matrix)
-        return self._eigenvectors
-
-    def exp(self, a):
-        """Computes a tensor corresponding to exp(-1j * a * H).
-
-        Args:
-            a (complex): Complex number to multiply Hamiltonian before
-                exponentiation.
-        """
-        if self._exp["a"] != a:
-            self._exp["a"] = a
-            if self._eigenvectors is None:
-                self._exp["result"] = K.linalg.expm(-1j * a * self.matrix)
-            else:
-                expd = K.linalg.diag(K.exp(-1j * a * self._eigenvalues))
-                ud = K.transpose(K.math.conj(self._eigenvectors))
-                self._exp["result"] = K.matmul(self._eigenvectors,
-                                               K.matmul(expd, ud))
-        return self._exp["result"]
-
-    def expectation(self, state, normalize=False):
-        """Computes the real expectation value for a given state.
-
-        Args:
-            state (array): the expectation state.
-            normalize (bool): If ``True`` the expectation value is divided
-                with the state's norm squared.
-
-        Returns:
-            Real number corresponding to the expectation value.
-        """
-        statec = K.math.conj(state)
-        hstate = self @ state
-        ev = K.math.real(K.reduce_sum(statec * hstate))
-        if normalize:
-            norm = K.reduce_sum(K.square(K.abs(state)))
-            return ev / norm
-        return ev
-
-    def __add__(self, o):
-        """Add operator."""
-        if isinstance(o, self.__class__):
-            if self.nqubits != o.nqubits:
-                raise RuntimeError('Only hamiltonians with the same '
-                                   'number of qubits can be added.')
-            new_matrix = self.matrix + o.matrix
-            return self.__class__(self.nqubits, new_matrix)
-        elif isinstance(o, NUMERIC_TYPES):
-            new_matrix = (self.matrix +
-                          o * K.eye(2 ** self.nqubits, dtype=self.matrix.dtype))
-            return self.__class__(self.nqubits, new_matrix)
-        else:
-            raise NotImplementedError(f'Hamiltonian addition to {type(o)} '
-                                      'not implemented.')
-
-    def __radd__(self, o): # pragma: no cover
-        """Right operator addition."""
-        return self.__add__(o)
-
-    def __sub__(self, o):
-        """Subtraction operator."""
-        if isinstance(o, self.__class__):
-            if self.nqubits != o.nqubits:
-                raise RuntimeError('Only hamiltonians with the same '
-                                   'number of qubits can be added.')
-            new_matrix = self.matrix - o.matrix
-            return self.__class__(self.nqubits, new_matrix)
-        elif isinstance(o, NUMERIC_TYPES): # pragma: no cover
-            new_matrix = (self.matrix -
-                          o * K.eye(2 ** self.nqubits, dtype=self.matrix.dtype))
-            return self.__class__(self.nqubits, new_matrix)
-        else:
-            raise NotImplementedError(f'Hamiltonian subtraction to {type(o)} '
-                                      'not implemented.')
-
-    def __rsub__(self, o):
-        """Right subtraction operator."""
-        if isinstance(o, self.__class__): # pragma: no cover
-            if self.nqubits != o.nqubits:
-                raise RuntimeError('Only hamiltonians with the same '
-                                   'number of qubits can be added.')
-            new_matrix = o.matrix - self.matrix
-            return self.__class__(self.nqubits, new_matrix)
-        elif isinstance(o, NUMERIC_TYPES):
-            new_matrix = (o * K.eye(2 ** self.nqubits, dtype=self.matrix.dtype)
-                          - self.matrix)
-            return self.__class__(self.nqubits, new_matrix)
-        else:
-            raise NotImplementedError(f'Hamiltonian subtraction to {type(o)} '
-                                      'not implemented.')
-
-    def __mul__(self, o):
-        """Multiplication to scalar operator."""
-        if isinstance(o, NUMERIC_TYPES):
-            new_matrix = self.matrix * o
-            r = self.__class__(self.nqubits, new_matrix)
-            if self._eigenvalues is not None:
-                if o.real >= 0:
-                    r._eigenvalues = o * self._eigenvalues
-                else:
-                    r._eigenvalues = o * self._eigenvalues[::-1]
-            if self._eigenvectors is not None:
-                if o.real > 0:
-                    r._eigenvectors = self._eigenvectors
-                elif o == 0:
-                    r._eigenvectors = K.eye(
-                        self._eigenvectors.shape[0], dtype=self.matrix.dtype)
-            return r
-        elif isinstance(o, (np.ndarray, K.Tensor)):
-            new_matrix = self.matrix * K.cast(o, dtype=self.matrix.dtype)
-            return self.__class__(self.nqubits, new_matrix)
-        else:
-            raise NotImplementedError(f'Hamiltonian multiplication to {type(o)} '
-                                      'not implemented.')
-
-    def __rmul__(self, o):
-        """Right scalar multiplication."""
-        return self.__mul__(o)
-
-    def __matmul__(self, o):
-        """Matrix multiplication with other Hamiltonians or state vectors."""
-        if isinstance(o, self.__class__):
-            new_matrix = K.matmul(self.matrix, o.matrix)
-            return self.__class__(self.nqubits, new_matrix)
-        elif isinstance(o, (K.Tensor, np.ndarray)):
-            rank = len(tuple(o.shape))
-            if rank == 1: # vector
-                return K.matmul(self.matrix, o[:, K.newaxis])[:, 0]
-            elif rank == 2: # matrix
-                return K.matmul(self.matrix, o)
-            else:
-                raise ValueError(f'Cannot multiply Hamiltonian with '
-                                  'rank-{rank} tensor.')
-        else:
-            raise NotImplementedError(f'Hamiltonian matrix multiplication to '
-                                       '{type(o)} not implemented.')
+from qibo.config import BACKEND_NAME, DTYPES
+if BACKEND_NAME == "tensorflow":
+    from qibo.tensorflow.hamiltonians import TensorflowHamiltonian as Hamiltonian
+else: # pragma: no cover
+    raise NotImplementedError("Only Tensorflow backend is implemented.")
 
 
 def _multikron(matrix_list):
